- Chaged parameters type in Power class from int to float. - Changed SS to CE in SPI class. - Added seed data for SBC. - Added migration for SBC.

+++ Entities/InputVoltage.cs
@@ -2,6 +2,6 @@
 
 public record InputVoltage
 {
-    public int Min { get; init; }
-    public int Max { get; init; }
+    public float Min { get; init; }
+    public float Max { get; init; }
 }
