+++ Entities/Power.cs
@@ -4,9 +4,9 @@ namespace Valkyrie.Entities;
 
 public record Power
 {
-    public int IOVoltage { get; init; }
+    public float IOVoltage { get; init; }
     public InputVoltage InputVoltage { get; init; } = new InputVoltage();
-    public int IOPinCurrent { get; init; }
+    public float IOPinCurrent { get; init; }
     public string PowerSupplyConnector { get; init; } = string.Empty;
